% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignprobandkappas.R
\name{assignprobandkappas}
\alias{assignprobandkappas}
\title{Computes probabilities of (relabeled) cluster and kappas.}
\usage{
assignprobandkappas(variables, k, metriccent, data, initialcluster)
}
\arguments{
\item{variables}{internally provided by \code{summary.miclust}.}

\item{k}{internally provided by \code{summary.miclust}.}

\item{metriccent}{internally provided by \code{summary.miclust}.}

\item{data}{internally provided by \code{summary.miclust}.}

\item{initialcluster}{internally provided by \code{summary.miclust}.}
}
\value{
internal value to be used by \code{summary.miclust}.
}
\description{
\code{assignprobandkappas} returns a list with information on probabilities of
  cluster belonging and Cohen's kappas.
}
\keyword{internal}
