% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerdata.R
\name{centerdata}
\alias{centerdata}
\title{Center data.}
\usage{
centerdata(data)
}
\arguments{
\item{data}{internally provided by \code{standardizedata} function.}
}
\value{
internal value to be used by \code{standardizedata} function.
}
\description{
\code{centerdata} centers all variables at the mean.
}
\keyword{internal}
