% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid.R
\name{centroid}
\alias{centroid}
\title{Computes centroid.}
\usage{
centroid(data, cluster, centpos)
}
\arguments{
\item{data}{internally provided by \code{doclusterkmeans} function.}

\item{cluster}{internally provided by \code{doclusterkmeans} function.}

\item{centpos}{internally provided by \code{doclusterkmeans} function.}
}
\value{
internal value to be used by \code{doclusterkmeans} function.
}
\description{
\code{centroid} computes the centroid for each cluster (mean o median).
}
\keyword{internal}
