% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doclusterkmeans.R
\name{doclusterkmeans}
\alias{doclusterkmeans}
\title{Performs K-means clustering with optional variable selection.}
\usage{
doclusterkmeans(
  search,
  data,
  k,
  metriccent,
  inertiapower = 1,
  maxvars = NULL,
  centpos,
  initcl
)
}
\arguments{
\item{search}{internally provided by \code{miclust} function.}

\item{data}{internally provided by \code{miclust} function.}

\item{k}{internally provided by \code{miclust} function.}

\item{metriccent}{internally provided by \code{miclust} function.}

\item{inertiapower}{internally provided by \code{miclust} function.}

\item{maxvars}{internally provided by \code{miclust} function.}

\item{centpos}{internally provided by \code{miclust} function.}

\item{initcl}{internally provided by \code{miclust} function.}
}
\value{
internal value to be used by \code{miclust} function.
}
\description{
\code{doclusterkmeans} performs K-means clustering with optional variable
  selection.
}
\keyword{internal}
