% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doclusterkmeansbackward.R
\name{doclusterkmeansbackward}
\alias{doclusterkmeansbackward}
\title{Performs K-means with backward selection.}
\usage{
doclusterkmeansbackward(data, k, metriccent, inertiapower = 1, centpos, initcl)
}
\arguments{
\item{data}{internally provided by \code{doclusterkmeans} function.}

\item{k}{internally provided by \code{doclusterkmeans} function.}

\item{metriccent}{internally provided by \code{doclusterkmeans} function.}

\item{inertiapower}{internally provided by \code{doclusterkmeans} function.}

\item{centpos}{internally provided by \code{doclusterkmeans} function.}

\item{initcl}{internally provided by \code{doclusterkmeans} function.}
}
\value{
internal value to be used by \code{doclusterkmeans} function.
}
\description{
\code{doclusterkmeansbackward} performs K-means clustering with backward
  variable selection.
}
\keyword{internal}
