% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doclusterkmeansnone.R
\name{doclusterkmeansnone}
\alias{doclusterkmeansnone}
\title{Performs K-means without variable selection.}
\usage{
doclusterkmeansnone(data, k, metriccent, inertiapower = 1, centpos, initcl)
}
\arguments{
\item{data}{internally provided by \code{doclusterkmeans} function.}

\item{k}{internally provided by \code{doclusterkmeans} function.}

\item{metriccent}{internally provided by \code{doclusterkmeans} function.}

\item{inertiapower}{internally provided by \code{doclusterkmeans} function.}

\item{centpos}{internally provided by \code{doclusterkmeans} function.}

\item{initcl}{internally provided by \code{doclusterkmeans} function.}
}
\value{
internal value to be used by \code{doclusterkmeans} function.
}
\description{
\code{doclusterkmeansnone} performs K-means clustering without variable
  selection.
}
\keyword{internal}
