% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Computes euclidean distance.}
\usage{
euclidean(x, centers)
}
\arguments{
\item{x}{internally provided by \code{miclust} function.}

\item{centers}{internally provided by \code{miclust} function.}
}
\value{
internal value to be used by \code{miclust} function.
}
\description{
\code{euclidean} computes the euclidean distance from data to centers.
}
\keyword{internal}
