% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcritcfkcca.R
\name{getcritcfkcca}
\alias{getcritcfkcca}
\title{Computes CritCF.}
\usage{
getcritcfkcca(kmeansfitted, inertiapower)
}
\arguments{
\item{kmeansfitted}{internally provided by \code{doclusterkmeans} function.}

\item{inertiapower}{internally provided by \code{doclusterkmeans} function.}
}
\value{
internal value to be used by \code{doclusterkmeans} function.
}
\description{
\code{getcritcfkcca} computes CritCF (see references).
}
\keyword{internal}
