% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getinitialcentroids.R
\name{getinitialcentroids}
\alias{getinitialcentroids}
\title{Computes initial centroids.}
\usage{
getinitialcentroids(data, ncentr)
}
\arguments{
\item{data}{internally provided by \code{doclusterkmeans} function.}

\item{ncentr}{internally provided by \code{doclusterkmeans} function.}
}
\value{
internal value to be used by \code{doclusterkmeans} function.
}
\description{
\code{getinitialcentroids} computes initial centroids for the clustering
  process.
}
\keyword{internal}
