% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{Computes Manhattan distance.}
\usage{
manhattan(x, centers)
}
\arguments{
\item{x}{internally provided by \code{miclust} function.}

\item{centers}{internally provided by \code{miclust} function.}
}
\value{
internal value to be used by \code{miclust} function.
}
\description{
\code{manhattan} computes the Manhattan distance from data to centers.
}
\keyword{internal}
