% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.miclust.R
\name{plot.miclust}
\alias{plot.miclust}
\title{Shows a graphical representation of the results.}
\usage{
\method{plot}{miclust}(x, k = NULL, ...)
}
\arguments{
\item{x}{object of class \code{miclust} obtained with the function \code{\link{miclust}}.}

\item{k}{number of clusters. The default value is the optimal number of clusters
obtained by \code{\link{miclust}}.}

\item{\dots}{further arguments for the plot function.}
}
\value{
a plot to visualize the clustering results.
}
\description{
Creates a graphical representation of the results of \code{\link{miclust}}.
}
\seealso{
\code{\link{miclust}}, \code{\link{summary.miclust}}.
}
