% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.miclust.R
\name{print.miclust}
\alias{print.miclust}
\title{Prints the results.}
\usage{
\method{print}{miclust}(x, ...)
}
\arguments{
\item{x}{object of class \code{miclust} obtained with the function \code{\link{miclust}}.}

\item{\dots}{further arguments for the print method.}
}
\value{
prints a description of the clustering main results.
}
\description{
Creates a summary print of the results of \code{\link{miclust}}.
}
