% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.miclust.R
\name{print.summary.miclust}
\alias{print.summary.miclust}
\title{Prints the summary of results.}
\usage{
\method{print}{summary.miclust}(x, digits = 2, ...)
}
\arguments{
\item{x}{object of class \code{summary.miclust} obtained with the method \code{\link{summary.miclust}}.}

\item{digits}{digits for the print method. Default is 2.}

\item{\dots}{further arguments for the print method.}
}
\value{
a print of the summary of the results generated by \code{\link{summary.miclust}}.
}
\description{
Prints the summary of the results of \code{\link{summary.miclust}}.
}
\seealso{
\code{\link{miclust}}, \code{\link{summary.miclust}}.
}
