% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabelclusters.R
\name{relabelclusters}
\alias{relabelclusters}
\title{Relabel clusters.}
\usage{
relabelclusters(refcluster, cluster)
}
\arguments{
\item{refcluster}{internally provided by \code{assignprobandkappas} function.}

\item{cluster}{internally provided by \code{assignprobandkappas} function.}
}
\value{
internal value to be used by \code{assignprobandkappas} function.
}
\description{
\code{relabelclusters} relabels the clusters so that they all have the same
  meaning in the all the data sets.
}
\keyword{internal}
