% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizedata.R
\name{standardizedata}
\alias{standardizedata}
\title{Standardize data.}
\usage{
standardizedata(data)
}
\arguments{
\item{data}{internally provided by \code{getdata} function.}
}
\value{
internal value to be used by \code{getdata} function.
}
\description{
\code{standardizedata} standardizes variables in data.
}
\keyword{internal}
